+++ apps/redi-talent-pool/src/components/organisms/JobseekerProfileCard.tsx
@@ -1,4 +1,4 @@
-import { NewProfileCard } from '@talent-connect/shared-atomic-design-components'
+import { ProfileCard } from '@talent-connect/shared-atomic-design-components'
 import {
   desiredPositionsIdToLabelMap,
   germanFederalStates,
@@ -45,7 +45,7 @@ export function JobseekerProfileCard({
 
   return (
     <div className="jobSeeker-profile-card-wrapper">
-      <NewProfileCard
+      <ProfileCard
         profile={{ id, avatar, fullName, location, languages }}
         subheader={subheader}
         tags={tags}
