+++ libs/shared-atomic-design-components/src/lib/molecules/ProfileCard.scss
@@ -5,7 +5,7 @@
   text-decoration: none;
 }
 
-.new-profile-card {
+.profile-card {
   display: flex;
   flex-direction: column;
   border-radius: 4px;
