+++ libs/shared-atomic-design-components/src/lib/molecules/index.ts
@@ -4,6 +4,6 @@ export { default as FaqItem } from './FaqItem'
 export { default as LightModal } from './LightModal'
 export { default as Modal } from './Modal'
 export { default as Module } from './Module'
-export { default as NewProfileCard } from './NewProfileCard'
 export { default as Pagination } from './Pagination'
 export { default as PipeList } from './PipeList'
+export { default as ProfileCard } from './ProfileCard'
